% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity-weighting.R
\name{add_propensity_weights}
\alias{add_propensity_weights}
\title{Calculates inverse probability of treatment weights using a random forest.}
\usage{
add_propensity_weights(df, treatment, ivs, impute_missing = FALSE)
}
\arguments{
\item{df}{<\code{tbl_df}> Data frame on which to perform weighting.}

\item{treatment}{<\code{tidy-select}> Treatment column to use in propensity score model.}

\item{ivs}{<\code{tidy-select}> Columns of independent variables to use in propensity score model.}

\item{impute_missing}{<\code{lgl}> If \code{TRUE} will internally impute missing values in IVs for propensity score fit.}
}
\value{
<\code{tbl_df}> Data frame with new column for propensity weights.
}
\description{
Calculates inverse probability of treatment weights using a random forest.
}
