% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbletest-package.R
\docType{data}
\name{example_dat}
\alias{example_dat}
\title{Example dataset for package demonstration}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 501 rows and 9 columns.
}
\usage{
example_dat
}
\description{
Example dataset for package demonstration
}
\keyword{datasets}
