% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_tbl.covariate_balance}
\alias{format_tbl.covariate_balance}
\title{Formats the descriptives table outputted by \code{covariate_balance}.}
\usage{
\method{format_tbl}{covariate_balance}(
  obj,
  add_Ns = TRUE,
  remove_duplicates = TRUE,
  remove_false = TRUE,
  std_diff_digits = 1L
)
}
\arguments{
\item{obj}{<\code{tbl_df}> Tibble outputted by \code{descriptives}.}

\item{add_Ns}{<\code{logical(1)}> Should treatment counts be added to column names?}

\item{remove_duplicates}{<\code{logical(1)}> Should repeated variable names and p-values for categorical variables be removed?
If TRUE, variable names and p-values will be presented only once on their own row.}

\item{remove_false}{<\code{logical(1)}> Should binary variables only show the N and percent of TRUE?}

\item{std_diff_digits}{<\code{integer(1)}> Number of digits to round standardized differences.}
}
\description{
Formats the descriptives table outputted by \code{covariate_balance}.
}
