% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_tbl.descriptives}
\alias{format_tbl.descriptives}
\title{Formats the descriptives table outputted by \code{descriptives}.}
\usage{
\method{format_tbl}{descriptives}(
  obj,
  add_Ns = TRUE,
  remove_duplicates = TRUE,
  remove_false = TRUE,
  p_val_digits = 3L
)
}
\arguments{
\item{obj}{<\code{tbl_df}> Tibble outputted by \code{descriptives}.}

\item{add_Ns}{<\code{logical(1)}> Should treatment counts be added to column names?}

\item{remove_duplicates}{<\code{logical(1)}> Should repeated variable names and p-values for categorical variables be removed?
If TRUE, variable names and p-values will be presented only once on their own row.}

\item{remove_false}{<\code{logical(1)}> Should binary variables only show the N and percent of TRUE?}

\item{p_val_digits}{<\code{integer(1)}> Number of digits to round p-values.}
}
\description{
Formats the descriptives table outputted by \code{descriptives}.
}
