% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_tbl}
\alias{format_tbl}
\title{Formats the descriptives table outputted by \code{descriptives}.}
\usage{
format_tbl(obj, ...)
}
\arguments{
\item{obj}{<\code{tbl_df}> Tibble outputted by \code{descriptives}.}

\item{add_Ns}{<\code{logical(1)}> Should treatment counts be added to column names?}

\item{remove_duplicates}{<\code{logical(1)}> Should repeated variable names and p-values for categorical variables be removed?
If TRUE, variable names and p-values will be presented only once on their own row.}

\item{remove_false}{<\code{logical(1)}> Should binary variables only show the N and percent of TRUE?}
}
\value{
<\code{tbl_df}> Formatted tibble
}
\description{
Allows for easy formatting of summary statistics table. Treatment group counts are added to the column names,
duplicated categorical variable names and p-values are removed, and p-values are rounded.
}
