% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{love_plot}
\alias{love_plot}
\title{Creates a Love plot comparing standardized differences before and after weighting}
\usage{
love_plot(unadj, adj)
}
\arguments{
\item{unadj}{<\code{tbl_df}> Tibble outputted by \code{covariate_balance()} without weighting.}

\item{adj}{<\code{tbl_df}> Tibble outputted by \code{covariate_balance()} after propensity score weighting.}
}
\value{
<\code{ggplot}> ggplot2 Love plot
}
\description{
Creates a Love plot comparing standardized differences before and after weighting
}
