% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocesses a tibble to prepare it for presentation in a descriptives table}
\usage{
preprocess(df, var_labels = NULL, clean_names = TRUE,
  convert_to_logical = TRUE, quiet = FALSE)
}
\arguments{
\item{df}{<\code{tbl_df}> Data frame with treatment and variables of interest as columns.}

\item{var_labels}{<\code{list}> Named list of column names and descriptive labels. By default, labels will generated by converting column names to title case}

\item{clean_names}{<\code{logical}> Should variable names be cleaned with \code{janitor::clean_names()}?}

\item{convert_to_logical}{<\code{logical}> Should binary variables be converted to logical?}

\item{quiet}{<\code{logical}> Should messages related to preprocessing changes be suppressed?}
}
\value{
<\code{tbl_df}>
}
\description{
Cleans the variable names, adds variable labels, and converts binary variables to logical in the output tibble
}
