% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity-weighting.R
\name{propensity_weighting}
\alias{propensity_weighting}
\title{Calculates inverse probability of treatment weights using a random forest}
\usage{
propensity_weighting(df, treatment, ivs)
}
\arguments{
\item{df}{<\code{tbl}> Data frame on which to perform weighting}

\item{treatment}{<\code{character(1)}> String with name of treatment column}

\item{ivs}{<\code{character}> Character vector with names of independent variabes to use in propensity score model}
}
\value{
<\code{tbl}> Data frame with new column for propensity weights
}
\description{
Calculates inverse probability of treatment weights using a random forest
}
