% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbletest-package.R
\docType{package}
\name{tibbletest-package}
\alias{tibbletest}
\alias{tibbletest-package}
\title{tibbletest: Clean Descriptives Tables with Significance Tests by Group}
\description{
A tool to easily create a table of descriptive statitics split by treatment group.
}
\author{
\strong{Maintainer}: Michael Battaglia \email{mbattagl@methodsconsultants.com}

Authors:
\itemize{
  \item Caleb Scheidel \email{caleb@methodsconsultants.com}
}

}
\keyword{internal}
