% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity-weighting.R
\name{add_propensity_weights}
\alias{add_propensity_weights}
\title{Calculates inverse probability of treatment weights (IPTW) using a random
forest}
\usage{
add_propensity_weights(df, treatment, ivs, impute_missing = FALSE)
}
\arguments{
\item{df}{<\code{tbl_df}> Data frame on which to perform weighting.}

\item{treatment}{<\code{tidy-select}> Treatment column to use in propensity score
model.}

\item{ivs}{<\code{tidy-select}> Columns of independent variables to use in
propensity score model.}

\item{impute_missing}{<\code{lgl}> If \code{TRUE} will internally impute missing values
in IVs for propensity score fit. Uses kNN with k = 5 for imputation.}
}
\value{
<\code{tbl_df}> Data frame with new column for propensity weights.
}
\description{
\code{add_propensity_weights()} adds a column of propensity weights to a data
frame. It calculates the propensity score using a random forest model to
predict the probability of treatment selection conditional on baseline
covariates. Stabilized inverse propensity weights are then calculated from
the propensity scores and these weights are returned as a new column in the
data frame.
}
\section{Random Forest Model}{
 A propensity score is the probability of a unit
being assigned to a particular treatment group given a set of baseline
characteristics. To calculate the propensity score for each unit, we fit a
model with treatment group as the outcome and baseline characteristics as
independent variables. The \code{ranger} package is used to fit a random
forest model to calculate the propensity scores. A probability forest is
fit by specifying \code{probability = TRUE}, and the
\code{respect.unordered.factors} argument is set to \code{TRUE}. All other
arguments are left at their defaults. See \code{\link[ranger]{ranger}} for
more details.
}

\section{Missing Data Imputation}{
 \code{ranger} does not support missing
values, so they have to be dealt with prior to running the model. By
setting \code{impute_missing = TRUE} the function will automatically impute
the independent variables prior to fitting the random forest model. It uses
k-nearest neighbors with \code{k = 5} for imputation.
}

\section{Stabilized Inverse Probability of Treatment Weights}{
 The IPTW for
each subject is calculated by taking the inverse of the probability that
they received the treatment that they actually received. That is, if a
subject is in the "control" group, then their weight is calculated as
\eqn{1 / P(Treatment = Control)}. An issue with traditional IPTWs is that
the sample size of the weighted data (sum of the weights) is higher than
the original sample size. An improvement to the traditional weights are
stabilized weights, which insure that the mean weight in each treatment
group is 1. The weights that the function outputs are stabilized weights.
}

