% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{Creates a table of summary statistics}
\usage{
descriptives(df, treatment = NULL, variables = NULL, weights = NULL)
}
\arguments{
\item{df}{<\code{tbl_df}> Data frame with treatment and variables of interest as columns.}

\item{treatment}{<\code{character(1)}> String with name of treatment column. If left blank, will produce univariate summary statistics.}

\item{variables}{<\code{character}> Character vector of column names. If left blank, will be inferred from data.}

\item{weights}{<\code{character(1)}> String with name of column with observation weights. If left blank, will not use observation weights.}
}
\value{
<\code{tbl_df}> Tibble with summary statistics split by treatment (optional).
}
\description{
Generates a table with summary statistics for each variable, optionally split by a grouping variable.
}
