% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_tbl}
\alias{format_tbl}
\title{Formats the descriptives table outputted by \code{descriptives}}
\usage{
format_tbl(obj, p_val_digits = 3L)
}
\arguments{
\item{obj}{<\code{tbl_df}> Dataframe outputted by \code{descriptives}}

\item{p_val_digits}{<\code{integer(1)}> Number of digits to round p-values}
}
\value{
<\code{tbl_df}> Formatted tibble
}
\description{
For categorical variables with multiple labels, variable name and p-value are moved to their own row and are not repeated. Additionally, p-values are rounded.
}
